# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users/{USER_ID}/memorials": {
            "get": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Get user memorials",
                "description": "Get user memorials",
                "operationId": "get-memorials",
                "parameters": [
                    {
                        "name": "USER_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ShortMemorialResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/memorials/{MEMORIAL_ID}": {
            "get": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Get memorial info",
                "description": "Get memorial info",
                "operationId": "get-memorial",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortMemorialResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Update memorial",
                "description": "Update memorial",
                "operationId": "update-memorial",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortMemorialResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Change memorial status",
                "description": "Change memorial status",
                "operationId": "change-memorial-status",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/memorials": {
            "post": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Create memorial",
                "description": "Create memorial",
                "operationId": "create-memorial",
                "parameters": [
                    {
                        "name": "default",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/api/v1/memorial/{MEMORIAL_ID}": {
            "delete": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Delete memorial",
                "description": "Delete all memorial",
                "operationId": "destroy-memorial",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/memorials/{MEMORIAL_ID}/avatar": {
            "post": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Upload memorial avatar",
                "description": "Upload memorial avatar",
                "operationId": "upload-memorial-avatar",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "file",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
...
        "/api/v1/users/{ID}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get full user details",
                "description": "Get full user details",
                "operationId": "details",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 11
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Vladimir Putin"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "http://b-ouquet.com/storage/avatars/3gpbiSkdpCbyIADiyb0wuopeSVnbT0gBCJfZi0GL_avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "update-user",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Disable user",
                "description": "Disable user",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ShortMemorialResource": {
                "properties": {
                    "id": {
                        "type": "int8",
                        "example": "53"
                    },
                    "title": {
                        "type": "string",
                        "example": "Lorem ipsum"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "Fitzgerald"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "place_of_birth": {
                        "type": "string",
                        "example": "New York City"
                    },
                    "place_of_death": {
                        "type": "string",
                        "example": "Concord"
                    },
                    "day_of_birth": {
                        "type": "string",
                        "example": "2012-12-12T00:00:00.000000Z"
                    },
                    "day_of_death": {
                        "type": "string",
                        "example": "2012-12-13T00:00:00.000000Z"
                    },
                    "text": {
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla vel libero leo."
                    },
                    "default": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "status": {
                        "type": "string",
                        "example": "published"
                    },
                    "user": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShortUserResource"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShortMemorialResource"
                }
            },
...
            "ShortUserResource": {
                "properties": {
                    "id": {
                        "type": "int8",
                        "example": "53"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "http://b-ouquet.com/storage/avatars/3gpbiSkdpCbyIADiyb0wuopeSVnbT0gBCJfZi0GL_avatar.jpg"
                    },
                    "memorial": {
                        "type": "int8",
                        "example": "35"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShortUserResource"
                }
            }
        },
...
    }
}